Optimise data format, reduced from 13 bytes to 5 bytes (including servo data). Note that servo logging is not yet implemented.

+++ Datastore.h
@@ -43,9 +43,9 @@ class LogEntry
     void setBattery(float battery);
     void setServo(uint8_t servo);
   private:
-    int32_t pressureRaw;
-    int32_t temperatureRaw;
-    float batteryRaw;
+    int16_t pressureRaw;
+    uint8_t temperatureRaw;
+    uint8_t batteryRaw;
     uint8_t servoRaw;
 } __attribute__ ((__packed__)); // this is to force the compiler not to pad the data structure. It probably makes no difference on AVR-GCC.
